added adaptive design, completed at 85%
